# perf(format): reuse matches instead of created when replacing