@iamkun 2019-02-02
:tada: This PR is included in version 1.8.1 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/dayjs)
- [GitHub release](https://github.com/iamkun/dayjs/releases/tag/v1.8.1)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: